Update TypeFinder.cpp
Figured out bug with adding the PkmnType objects to a map.
Added function to take user input for testing combos.
Added a findMostResistances() function that finds the combination that has the most resistances, with a given type. (Doesn't include the combination of itself).
Added two helper functions for printing 1D and 2D vectors.
Gave each type object a name attribute for easier identification using only the object. Could probably remove a the use of a map now, though a map does allow for quicker access to each type.