Update TypeFinder.cpp
Coded returnAll() function to instantly find all the type combinations with 4 resistances or 8 strengths. Doesn't work half the time I ran it on SoloLearn - not sure where the bug is. Haven't checked. 

There was a removeRepeats() function I coded, to remove repeat combinations in the returnAll() function, but then I realized it'd do nothing since the type combos are "type1-type2", so Water.getCombo(Dark)'s name will be "Water-Dark", but Dark.getCombo(Water) will be "Dark-Water". For each combination I'd need to have a way to separate the text at either side of the hyphen ("Water-Dark" into {"Water", "Dark"}. Then for run through the rest of the list, split each combo as well, and check if their contents are the same.
Working on function to get input from the user and automatically call either the findMostResistances(), findMostStrengths, findLeastWeaknesses(), or returnAll() functions based on a string they type in.