Update TypeFinder.cpp
Changed the arguments of returnAll() I was testing in the main function. The returnAll() function still has a bug of often not working due to something with invoking the function. As well as, if a number and interaction are passed that don't exist, (such as trying to find a type combo that has exactly 18 weaknesses) the function will stop the code from running-"no output".
I'm going to stop working on the project for now-I gave myself a set end date. I got most of what I wanted done but didn't fully implement the user input function.
I need to meet my deadlines and define what I want to achieve in the beginning, so there isn't this lukewarm feeling.